Добавляет модульную структуру
